pulled connect menu out of topbar and into separate component
